Add `(un)shuffle_(buf|into)`
The `shuffle_buf` and `unshuffle_buf` functions are like `shuffle` and
`unshuffle`, but allow the caller to control allocation.  And they are
very important for our benchmark; without them we're really just
benchmarking Vec::with_capacity().

`shuffle_into` and `unshuffle_into` also allow the caller to control
allocation, but they technically require the buffer to be initialized,
where as `shuffle_buf` and `unshuffle_buf` do not.